Remove requirementService and userService files, refactor imports in AccountSettings and Clearance components to use updated service paths. Clean up AuthContext by removing commented code.
